requirements.txt
pytesseract
pdf2image
pillow
pandas
